Calculator update: seems to check out more accurately
Calculations work out, it ignores part of what maxme was going for (I
think) by attempting to match as many transactions down the individual
level. This seems incredibly risky becuase it seems to assume we will
be able to beat all other simulatenous orders to the market. Fees are
now also integrated as inherited from the individual public market
modules and give a much more realistic view of actual returns.